bpf: nodeport: split off the egress-specific parts
This picks up a suggestion by Timo in
https://github.com/cilium/cilium/pull/30334#discussion_r1459344032
to split the big nodeport.h file into ingress/egress specific parts, so
that they can be included separately. This helps the XDP program, where
we don't need any of the egress code - but currently still compile the
tailcalls and then prune them out at load time. Having ingress/egress
parts also makes it much easier to understand the code flow.

To reduce the churn for now this patch only isolates the egress parts,
while nodeport.h contains the ingress *and* shared parts.
